Back-end/CRUD_API/requirements.txt
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.78.0
firebase==3.0.1
firebase-admin==6.0.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.4.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.14.0
httplib2==0.20.4
idna==3.4
msgpack==1.0.4
numpy==1.23.3
pandas==1.4.2
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.4
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.17.6
